Fixed issue with region not importing
