+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "valorant.js",
-  "version": "1.4.1",
+  "version": "1.4.3",
   "description": "API Wrapper for valorant with oauth support",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
